feat: implement native tooltip functionality for COG layers and refactor tooltip handling
